requirements.txt
Django==2.1.7
canonicalwebteam.http==1.0.1
canonicalwebteam.blog==1.0.1
canonicalwebteam.versioned-static==1.0.2
canonicalwebteam.yaml-responses[django]==1.1.0
canonicalwebteam.django_views==1.3.4
django-asset-server-url==0.1
django-template-finder-view==0.3
whitenoise==4.1.2
talisker==0.11.1
gunicorn[gevent]
